requirements.txt
celery==5.3.4
Django==4.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-timezone-field==6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
mysqlclient==2.2.0
pandas==2.1.0
python-decouple==3.8
redis==5.0.0
typing_extensions==4.7.1
whitenoise==6.5.0
